add organization to the factories module
