+++ newsfragments/938.misc.rst
@@ -1,0 +1,1 @@
+Code re-organization of ``p2p.tools.factories``
