+++ p2p/tools/factories/cancel_token.py
@@ -1,0 +1,10 @@
+import factory
+
+from cancel_token import CancelToken
+
+
+class CancelTokenFactory(factory.Factory):
+    class Meta:
+        model = CancelToken
+
+    name = factory.Sequence(lambda n: "test-token-{}".format(n))
