+++ p2p/tools/factories/socket.py
@@ -1,0 +1,10 @@
+import socket
+
+
+def get_open_port() -> int:
+    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    s.bind(("", 0))
+    s.listen(1)
+    port = s.getsockname()[1]
+    s.close()
+    return port
